separated navigator actions
